requirements-test.txt
--requirement requirements.txt
faker
pytest
pytest-cov
pytest-django